DRYD-781: Fix field option lists being cleared when a group condition is removed before a group is selected.

+++ src/components/search/input/GroupConditionInput.jsx
@@ -116,7 +116,7 @@ export default class GroupConditionInput extends Component {
 
     const path = condition.get('path');
 
-    if (deleteOptionList) {
+    if (path && deleteOptionList) {
       deleteOptionList(getRecordFieldOptionListName(recordType, path));
       deleteOptionList(getRecordGroupOptionListName(recordType, path));
     }
